requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-debug-toolbar==2.2
django-extensions==3.0.8
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
